move independent varibles to separate helper file
